fpe
Module ff1

A Rust implementation of the FF1 algorithm, specified in NIST Special Publication 800-38G.

Structs

BinaryNumeralString

A numeral string with radix 2.

FF1

A struct for performing FF1 encryption and decryption operations.

FlexibleNumeralString

A numeral string that supports radixes in [2..2^16).

InvalidRadix

Error indicating that a radix was not in the supported range of values for FF1.

Enums

NumeralStringError

Errors that can occur while using FF1 for encryption or decryption.

Traits

NumeralString

For a given base, a finite, ordered sequence of numerals for the base.

Operations

Type representing FF1 operations that can be performed on a sub-section of a NumeralString.