fpe::ff1
Struct FF1

pub struct FF1<CIPH: BlockCipher> { /* private fields */ }

A struct for performing FF1 encryption and decryption operations.

Implementations

impl<CIPH: BlockCipher + KeyInit> FF1<CIPH>

pub fn new(key: &[u8], radix: u32) -> Result<Self, InvalidRadix>

Creates a new FF1 object for the given key and radix.

Returns an error if the given radix is not in [2..2^16].

impl<CIPH: BlockCipher + BlockEncrypt + Clone> FF1<CIPH>

pub fn encrypt<NS: NumeralString>( &self, tweak: &[u8], x: &NS, ) -> Result<NS, NumeralStringError>

Encrypts the given numeral string.

Returns an error if the numeral string is not in the required radix.

pub fn decrypt<NS: NumeralString>( &self, tweak: &[u8], x: &NS, ) -> Result<NS, NumeralStringError>

Decrypts the given numeral string.

Returns an error if the numeral string is not in the required radix.