fpe::ff1
Trait Operations

pub trait Operations: Sized {
    type Bytes: AsRef<[u8]>;

    // Required methods
    fn numeral_count(&self) -> usize;
    fn to_be_bytes(&self, radix: u32, b: usize) -> Self::Bytes;
    fn add_mod_exp(
        self,
        other: impl Iterator<Item = u8>,
        radix: u32,
        m: usize,
    ) -> Self;
    fn sub_mod_exp(
        self,
        other: impl Iterator<Item = u8>,
        radix: u32,
        m: usize,
    ) -> Self;
}

Type representing FF1 operations that can be performed on a sub-section of a NumeralString.

Required Associated Types

type Bytes: AsRef<[u8]>

Type used for byte representations.

Required Methods

fn numeral_count(&self) -> usize

Returns the number of numerals in this numeral sub-string.

fn to_be_bytes(&self, radix: u32, b: usize) -> Self::Bytes

Returns a b-byte big-endian representation of the number that is encoded in base radix by this numeral string. The numerals are valued in decreasing order of significance (big-endian order). This corresponds to $STR^{b}{256}(NUM{radix}(X))$ in the NIST spec.

fn add_mod_exp( self, other: impl Iterator<Item = u8>, radix: u32, m: usize, ) -> Self

Computes (self + other) mod radix^m.

fn sub_mod_exp( self, other: impl Iterator<Item = u8>, radix: u32, m: usize, ) -> Self

Computes (self - other) mod radix^m.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.