fpe::ff1
Trait NumeralString

pub trait NumeralString: Sized {
    type Ops: Operations;

    // Required methods
    fn is_valid(&self, radix: u32) -> bool;
    fn numeral_count(&self) -> usize;
    fn split(&self) -> (Self::Ops, Self::Ops);
    fn concat(a: Self::Ops, b: Self::Ops) -> Self;
}

For a given base, a finite, ordered sequence of numerals for the base.

Required Associated Types

type Ops: Operations

Type used for FF1 computations.

This may be Self, or it may be a type that can operate more efficiently.

Required Methods

fn is_valid(&self, radix: u32) -> bool

Returns whether this numeral string is valid for the base radix.

fn numeral_count(&self) -> usize

Returns the number of numerals in this numeral string.

fn split(&self) -> (Self::Ops, Self::Ops)

Splits this numeral string of length n into two sections of lengths u = floor(n / 2) and v = n - u that can be used for FF1 computations.

fn concat(a: Self::Ops, b: Self::Ops) -> Self

Concatenates two strings used for FF1 computations into a single numeral string.

Object Safety

This trait is not object safe.